refactor: split runtime/98_global_scope.js
