+++ ext/node/polyfills/console.ts
@@ -4,7 +4,7 @@
 // deno-lint-ignore-file prefer-primordials
 
 import { Console } from "ext:deno_node/internal/console/constructor.mjs";
-import { windowOrWorkerGlobalScope } from "ext:runtime/98_global_scope.js";
+import { windowOrWorkerGlobalScope } from "ext:runtime/98_global_scope_shared.js";
 // Don't rely on global `console` because during bootstrapping, it is pointing
 // to native `console` object provided by V8.
 const console = windowOrWorkerGlobalScope.console.value;
